# [aws-rds] Make it possible to get normal value from all engine.