@itchyny 2017-01-23
The labels like OldestReplicationSlotLag seem to be kept as the same but I think more readable labels are preferable... (you can change in this pull request or create a new one, I leave it to you)