@sioncojp 2017-01-23
I'm gonna fix that [L100](https://github.com/mackerelio/mackerel-agent-plugins/blob/master/mackerel-plugin-aws-rds/lib/metrics.go#L100) (differece to name, label)