@sioncojp 2017-01-23
@itchyny 
Please give me some advice that I think some label names like:
```
{Name: "OldestReplicationSlotLag", Label: "SlotLag"},
{Name: "ReplicationSlotDiskUsage", Label: "SlotDiskUsage"},
{Name: "AuroraReplicaLagMaximum", Label: "Maximum"},
```
If you will approve, I'm gonna fix like that.