@sioncojp 2017-01-23
Fix all. Please re-review :)